-as smartctl is too slow, turning the page rendering painfully slow, display the *previous* value if it is not older than 2 min. The previous value is calculated at the script end, after the page is rendered. -shows read-only or read-write filesystem status -replace a few long "echo" by here-docs -fix uptime display -fix MTU display -use <th> instead of <td> -show filesystem running operations -sends the page as it is built (slow rendering, fast turn-around) instead of caching it and sending it all at once (slow turn-around, fast rendering)

+++ customroot/usr/www/cgi-bin/status.cgi
@@ -17,53 +17,63 @@ md_standby() {
 disk() {
 	nm="$1"
 	dsk="$2"
-	temp="??"
+	temp="--"
 	fam=""
 	mod="None"
 	tcap=""
 	pstatus=""
-	health_st=""
+	health_st="--"
 
 	if test -b /dev/$dsk; then
 		mod=$(cat /sys/block/$dsk/device/model)
-		tcap=$(awk -v sz=$(cat /sys/block/$dsk/size) 'BEGIN{printf "%.1fG", sz*512/1e9}' /dev/null)
+		tcap=$(awk '{printf "%.1f GB", $1*0.5e-6}' /sys/block/$dsk/size)
 		pstatus="$(eval echo \$power_mode_$dsk)"
 		if test "$pstatus" != "standby"; then
-			# smartctl is too slow, +/- 0.5 sec
-			res=$(smartctl -iA /dev/$dsk)
-			if test $? = "0"; then
-				health_st="OK"
-			elif test $? = "1"; then
-				health_st="unknown"
-			else
-				health_st="<font color=red> Failing</font>"
+			# smartctl is too slow, +/- 0.5 sec, use the *previous* value if not older than 2 minutes
+			if test -f /tmp/SMART-$dsk; then
+				if test "$(expr $(date +%s) - $(stat -t /tmp/SMART-$dsk | cut -d" " -f 14))" -gt 120; then
+					res=$(smartctl -iA /dev/$dsk)
+					if test $? = "0"; then
+						health_st="OK"
+					elif test $? = "1"; then
+						health_st="unknown"
+					else
+						health_st="<font color=red> Failing</font>"
+					fi
+					eval $(echo "$res" | awk '/^194/ {printf "temp=\"%d\";", $10}
+						/^Model Family/ {printf "fam=\"%s\";", $3}
+						/^Device:/ {printf "fam=\"%s\";", $2}')
+				else
+					. /tmp/SMART-$dsk
+				fi
 			fi
-			eval $(echo "$res" | awk '/^194/ {printf "temp=\"%d\";", $10}
-				/^Model Family/ {printf "fam=\"%s\";", $3}
-				/^Device:/ {printf "fam=\"%s\";", $2}')
 		fi
 	fi	
 
-	echo "<tr align=center>
-			<td align=left> $s $nm $es </td>
-			<td align=left> $fam $mod</td>
-			<td> $tcap </td>
-			<td> $pstatus </td>
-			<td> $temp </td> 
-			<td> $health_st </td>
-			</tr>"
+	cat<<-EOF
+		<tr align=center>
+		<td align=left> $s $nm $es </td>
+		<td>$dsk</td>
+		<td align=left> $fam $mod</td>
+		<td> $tcap </td>
+		<td> $pstatus </td>
+		<td> $temp </td> 
+		<td> $health_st </td>
+		</tr>
+	EOF
 }
 
 filesys() {
 	dsk="$1"
-	type=$2
+	dev=$(basename $dsk)
+	type=$(blkid -s TYPE -o value $dsk)
 	lbl="$(plabel $dsk)"
-	if test -z "$lbl"; then lbl=$dsk; fi
+	#if test -z "$lbl"; then lbl=$dev; fi
 	
-	if test -b /dev/$dsk; then
+	if test -b $dsk; then
 		cnt=""; dirty="";
-		eval $(df -h /dev/$dsk | awk '/'$dsk'/{printf "cap=%s;free=%s", $2, $4}')
-		res=$(tune2fs -l /dev/$dsk 2> /dev/null)
+		eval $(df -h $dsk | awk '/'$dev'/{printf "cap=%s;free=%s", $2, $4}')
+		res=$(tune2fs -l $dsk 2> /dev/null)
 		if test $? = 0; then
 			eval $(echo "$res" | awk \
 				'/state:/{ if ($3 != "clean") printf "dirty=*;"} \
@@ -71,16 +81,27 @@ filesys() {
 				/Maximum/{FS=":"; max_cnt=$2} \
 				END{printf "cnt=%d", max_cnt-curr_cnt}')
 		fi
+		MD="$(awk '/'$dev'/{ n = split($4, a,",")
+			for (i=1;i<=n;i++)
+				if (a[i] == "ro")
+					printf ("<font color=red> %s </font>", toupper(a[i]))
+				else if(a[i] == "rw")
+					print toupper(a[i]) }' /proc/mounts)"
 	fi
 
 	if test "$cnt" -lt 5; then cnt="<font color=RED> $cnt </font>"; fi
-	echo "<tr align=center>
-		<td align=left>$lbl </td>
-		<td> $cap </td> <td>$free </td>
-		<td> $type </td>
-		<td> <font color=RED> $dirty </font> </td>
-		<td> $cnt </td>
-		</tr>"
+	cat<<-EOF
+		<tr align=center>
+		<td>$dev</td>
+		<td>$lbl</td>
+		<td align=right>${cap}B</td>
+		<td align=right>${free}B </td>
+		<td>$type</td>
+		<td>$MD</td>
+		<td><font color=RED>$dirty</font> </td>
+		<td>$cnt</td>
+		</tr>
+	EOF
 }
 
 main() {
@@ -99,26 +120,28 @@ s="<strong>"
 es="</strong>"
 
 # do this early, to avoid affecting the load value		
-eval $(uptime | awk '{i=3; msg=""; while ($i != "load") \
-	{ msg=msg " " $i; i++; } printf "out=\"%s\"; load=%s", msg, $(i+2)}')
+eval $(awk '{ days = $1/86400; hours = $1 % 86400 / 3600; \
+	printf "up=\"%d day(s) %d hour(s)\"", days, hours }' /proc/uptime)
+
+load=$(cut -f1 -d" " /proc/loadavg)
 
 eval $(ethtool eth0 | awk '/Speed/{printf "Speed=%s", $2}')
 
 eval $(ifconfig eth0 | awk \
 		'/RX bytes/{printf "Rx=\"%s %s\";", $3, $4} \
 		/TX bytes/{printf "Tx=\"%s %s\";", $7, $8} \
-		/MTU/{printf "MTU=%s;", substr($5, 5)} \
+		/MTU/{printf "MTU=%s;", substr($0, match($a,"MTU")+4,5)} \
 		/HWaddr/{printf "MAC=\"%s\";", $5}' | tr "()" "  ")
 
 eval $(cat $temp_dev | awk '{printf "temp=\"%.1f\"", $1 / 1000 }')
 
-eval $(free | awk '/Swap/{printf "swap=\"%d/%d MB\"", $3/1024, $4/1024}')
+eval $(free | awk '/Swap/{printf "swap=\"%.1f/%d MB\"", $3/1024, $4/1024}')
 
 cat <<-EOF
 	<fieldset><Legend> $s System $es </legend>
-	$s Temperature:	 $es  $temp C
+	$s Temperature:	 $es  $temp
 	$s Fan speed:  $es $(cat $fan_dev) RPM 
-	<br> $s Uptime:	 $es $out $s Load: $es $load  
+	<br> $s Uptime:	 $es $up $s Load: $es $load  
 	$s Swap:  $es $swap 
 	</fieldset><br>
 
@@ -140,11 +163,12 @@ if test -z "$(ls /dev/sd?? 2>/dev/null)"; then
 else
 	cat<<-EOF
 		<tr align=center><td align=left> $s Bay $es </td>
-		<td> $s Model $es </td>
-		<td> $s Capacity $es </td>
-		<td> $s Power Status $es </td>
-		<td> $s Temp $es C </td>
-		<td> $s Health $es </td>
+		<th>Dev.</th>
+		<th>Model</th>
+		<th>Capacity</th>
+		<th>Power Status</th>
+		<th>Temp</th>
+		<th>Health</th>
 		</tr>
 	EOF
 
@@ -164,22 +188,24 @@ cat<<-EOF
 	<fieldset><legend> $s Mounted Filesystems $es </legend>
 EOF
 
-if test -z "$(awk '/^\/dev\/(sd|md)/{ print "yes"}' /proc/mounts)"; then
+if ! grep -q '^/dev/\(sd\|md\)' /proc/mounts; then
 	echo "None"
 else
 	cat<<-EOF
 		<table>
 		<tr align=center>
-		<td align=left> $s Label $es </td>
-		<td> $s Capacity $es</td> <td> $s Available $es</td>
-		<td> $s FS $es </td><td> $s Dirty $es </td> <td> $s FSCK $es </td> 
+		<th>Dev.</th>
+		<th>Label</th>
+		<th>Capacity</th><th>Available</th>
+		<th>FS</th><th>Mode</th>
+		<th>Dirty</th><th>FSCK</th> 
 		</tr>
 	EOF
 
-	while read dev mnt type rest; do
+	while read dev mnt rest; do
 		dsk=$(echo $dev | grep '^/dev/\(sd\|md\)')
 		if test -n "$dsk"; then
-			filesys $(basename $dsk) $type
+			filesys $dsk
 		fi
 	done < /proc/mounts
 fi
@@ -189,20 +215,20 @@ cat<<-EOF
 	<fieldset><Legend> $s RAID $es </legend><table>
 EOF
 
-#echo $check_md0, $check_md1, $check_md2
-
 if ! test -e /proc/mdstat; then
 	echo "None<br>"
 elif test -z "$(grep ^md /proc/mdstat)"; then
 	echo "None<br>"
 else
-	echo "<tr align=center>
-		<td align=left> $s Label $es </td> 
-		<td> $s Capacity $es</td>
-		<td> $s Type $es </td> <td> $s State $es </td>
-		<td> $s Status $es </td> <td> $s Action $es </td>
-		<td> $s Done $es </td> <td> $s ETA $es </td>
- 		</tr>"
+	cat<<-EOF
+		<tr align=center>
+		<th align=left>Dev.</th> 
+		<th>Capacity</th>
+		<th>Level</th><th>State</th>
+		<th>Status</th><th>Action</th>
+		<th>Done</th><th>ETA</th>
+ 		</tr>
+	EOF
 
 	for i in /dev/md[0-9]*; do
 		mdev=$(basename $i)
@@ -218,7 +244,8 @@ else
 		if test $? = 1; then lbl=$mdev; fi
 		sz=""; deg=""; act=""; compl=""; exp="";
 		if test "$state" != "inactive"; then
-			eval $(df -h | awk '/'$mdev'/{printf "sz=%s", $2}')
+			#eval $(df -h | awk '/'$mdev'/{printf "sz=%s", $2}')
+			sz=$(awk '{ printf "%.1f", $1/2/1024/1024}' /sys/block/$mdev/size)
 			if test "$type" = "raid1" -o "$type" = "raid5"; then
 				if test "$(cat /sys/block/$mdev/md/degraded)" != 0; then
 					deg="<font color=RED> degraded </font>"
@@ -227,31 +254,64 @@ else
 				fi
 				act=$(cat /sys/block/$mdev/md/sync_action)
 				if test "$act" != "idle"; then
+					act="<font color=RED> $act </font>"
 					compl=$(awk '{printf "%.1f%%", $1 * 100 / $3}' /sys/block/$mdev/md/sync_completed)
 					speed=$(cat /sys/block/$mdev/md/sync_speed)
 					exp=$(awk '{printf "%.1fmin", ($3 - $1) * 512 / 1000 / '$speed' / 60 }' /sys/block/$mdev/md/sync_completed)
 				fi
 			fi
 		fi
-		echo "<tr align=center>
+		cat<<-EOF
+			<tr align=center>
 			<td align=left>$lbl</td> 
-			<td>$sz</td>
-			<td>$type</td> <td>$state</td>
-			<td>$deg</td> <td>$act</td> <td>$compl</td> <td>$exp</td>
-			</tr>"
+			<td>${sz} GB</td>
+			<td>$type</td>
+			<td>$state</td>
+			<td>$deg</td>
+			<td>$act</td>
+			<td>$compl</td>
+			<td>$exp</td>
+			</tr>
+		EOF
 	done
 fi
 
-cat <<-EOF
-	</table></fieldset><br>
-	<fieldset><Legend> $s Printers $es </legend><table>
-EOF
+echo "</table></fieldset><br>"
+
+if test -n "$(ls /tmp/clean-* /tmp/format-* /tmp/convert-* 2>/dev/null)"; then
+
+	cat<<-EOF
+		<fieldset><Legend> $s Filesystem Maintenance $es </legend>
+		<table><tr><th>Part.</th><th>Label</th><th>Operation</th></tr>
+	EOF
+
+	for j in /dev/sd[a-z][1-9] /dev/md[0-9]*; do
+		part=$(basename $j)
+		for k in clean format convert; do
+			if test -f /tmp/$k-$part; then
+				if test -d /proc/$(cat /tmp/$k-$part.pid); then
+					cat<<-EOF
+						<tr><td>$part</td><td>$(plabel $part)</td>
+						<td><font color=RED>${k}ing...</font></td>
+						</tr>
+					EOF
+				else
+					rm -f /tmp/$k-$part*
+				fi
+			fi
+		done
+	done
+	echo "</table></fieldset>"
+fi
+
+echo "<fieldset><Legend> $s Printers $es </legend><table>"
 
 if ! test -f /etc/printcap; then
 	echo "None"
+elif ! grep -q '^[^#]' /etc/printcap; then
+	echo "None"
 else
-	echo "<tr align=center><td> $s Name $es</td>
-		<td> $s Model $es </td><td> $s Jobs $es</td></tr>"
+	echo "<tr align=center><th>Name</th><th>Model</th><th>Jobs</th></tr>"
 
 	while read ln; do
 		if test "${ln:0:1}" = "#"; then continue; fi
@@ -264,25 +324,18 @@ else
 		else
 			#host="remote"
 			if test -f /usr/bin/lpstat; then
-				jb=$(lpstat | grep ^$pr | wc -l)
+				jb=$(lpstat 2> /dev/null | grep ^$pr | wc -l)
 			fi
 		fi
 		echo "<tr><td>$pr</td><td>$desc</td><td>$jb</td></tr>"
 		
 	done < /etc/printcap
 fi
-echo "</table></fieldset><br>"
 
-#cat <<-EOF
-#	<fieldset><Legend> $s Software versions $es </legend>
-#	$s Alt-F: $es $(awk '/version/{print $2}' /etc/Alt-F)
-#	$s Linux: $es $(uname -r) 
-#	$s BusyBox: $es $(awk '/version:/ {print $4}' /boot/busybox.config) 
-#	$s Uclib: $es $(awk '/Version:/ {print $3}' /boot/uclibc.config) 
-#	</fieldset>
-#EOF
-
-echo "</body></html>"
+cat<<-EOF
+	</table></fieldset><br>
+	</body></html>
+EOF
 
 }
 
@@ -291,11 +344,36 @@ echo "</body></html>"
 # to avoid this, the page is generated to a temporary file and sent when
 # all is done. 
 # The time spent is the same, but not seeing the page being slowly
-# generated
-# avoids the slowness sensation
+# generated avoids the slowness sensation
+# hmmm, dont like it either.
 
-TF=/tmp/.status.tmp
+if false; then
+	TF=/tmp/.status.tmp
 
-main > $TF 2>&1
-cat $TF
-rm $TF >/dev/null 2>&1
+	main > $TF 2>&1
+	cat $TF
+	rm $TF >/dev/null 2>&1
+else
+	main
+fi
+
+# smartctl is too slow. Run it now and use the colected values in the *next* run
+for i in /dev/sd?; do
+	dsk=$(basename $i)
+	pstatus="$(eval echo \$power_mode_$dsk)"
+	if test "$pstatus" != "standby"; then
+		res=$(smartctl -iA $i)
+		if test $? = "0"; then
+			health_st="OK"
+		elif test $? = "1"; then
+			health_st="unknown"
+		else
+			health_st="<font color=red> Failing</font>"
+		fi
+		temp="??"; fam=""
+		eval $(echo "$res" | awk '/^194/ {printf "temp=\"%d\";", $10}
+			/^Model Family/ {printf "fam=\"%s\";", $3}
+			/^Device:/ {printf "fam=\"%s\";", $2}')
+	fi
+	echo "health_st=$health_st; temp=$temp; fam=$fam" > /tmp/SMART-$(basename $i)
+done
